# /active_series: add support for protobuf response format