--- a/pkg/querier/active_series_handler.go
+++ b/pkg/querier/active_series_handler.go
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+package querier