<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This adds protobuf as an alternative response encoding for the `/api/v1/cardinality/active_series` endpoint.
Clients can select whether they want to receive a response in json or protobuf format by setting the `Accept` header to the mime type of the desired encoding.

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/querier
                                                   │  bench.txt  │
                                                   │   sec/op    │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       980.6µ ± 1%
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10   84.15µ ± 4%

                                                   │  bench.txt   │
                                                   │     B/op     │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       963.6Ki ± 1%
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10   168.9Ki ± 1%

                                                   │  bench.txt  │
                                                   │  allocs/op  │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       13.15k ± 0%
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10    151.0 ± 0%
```

#### Which issue(s) this PR fixes or relates to

n/a

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`