@flxbk 2023-11-15
I have also run the benchmark added in this PR to compare the performance of protobuf marshaling in this PR to json marshaling based on json-iter, with the following results:
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/querier
                                                   │   stdjson    │              jsoniter              │
                                                   │    sec/op    │   sec/op     vs base               │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       990.59µ ± 5%   97.85µ ± 5%  -90.12% (p=0.002 n=6)
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10    85.91µ ± 2%   83.45µ ± 2%   -2.87% (p=0.004 n=6)
geomean                                               291.7µ        90.36µ       -69.02%

                                                   │   stdjson    │              jsoniter               │
                                                   │     B/op     │     B/op      vs base               │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       963.5Ki ± 1%   146.0Ki ± 1%  -84.85% (p=0.002 n=6)
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10   168.8Ki ± 1%   168.9Ki ± 1%        ~ (p=1.000 n=6)
geomean                                              403.3Ki        157.0Ki       -61.06%

                                                   │   stdjson    │              jsoniter               │
                                                   │  allocs/op   │ allocs/op   vs base                 │
ActiveSeriesHandler_ServeHTTP/json_encoding-10       13154.0 ± 0%   151.0 ± 0%  -98.85% (p=0.002 n=6)
ActiveSeriesHandler_ServeHTTP/protobuf_encoding-10     151.0 ± 0%   151.0 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                               1.409k        151.0       -89.29%
¹ all samples are equal
```

At least for the benchmark payload, both seem to be similarly performant, which I'm abandoning this PR for now.